Add patch from @jberends. Thanks !
Corrected a bug on multiple save of versioned documents saved with
an objectID() as '_id'.

The retrieval of the get_last_revision_id is now checked against
a unicode version of the objectID (as the unicode version of the
ObjectID is stored in the version container).

Also Added two tests to check for that and they both pass at the moment.