I've rewritten a few of the pages of documentation to make it clearer for complete noobies like myself to get started.
Specifically, I've modified the tutorial page to be more of a tutorial,
linking to other pages to provide more detail about concepts such as the
Document definition, structure, descriptors, and querying. Some of the content
that was in the tutorial, I made into a Document page. I also merged some
related pages (such as Polymorphism) into the Document page. I applied the
same merging to the Structure page which now includes the schemaless and custom
types definitions.

I also expanded to the structure page to make it more obvious about how to use
lists and dicts to create more complex structures.

I plan on tackling the Descriptors (and specifically, validation) next, but I
wanted to get this checked in first.

Finally, I applied a style guide to make the code examples more consistent
(probably missed a few spots on the first pass):

- I removed the >>>/... from 'class' definition snippets to make them easier to
  copy and paste into code. I left them for one line command examples.
- Everything modified should be limited to 80 character line lengths.
- All strings in code now use single quotes instead of double (was mixed
  before).
- Made examples more verbose where possible and avoided 'foo', 'bar', 'bla'
- Variables and code snippets now use `` `` everywhere so they're easier to spot
  in text

Minor bug fixes
- Upped version to 0.7/0.7.2 to represent the most current version of the docs
- Removed the 'show documentation source' link as it is often confused with the
  the 'show library source' desire